Initial implementation of a generic call tracking decorator.
This decorator checks that the repository in which the function was
initially defined is clean, then calls the function and adds an
annotated tag to the repository.

The tracking has to be activated in the config.

This is only an initial version, more configurability should be added,
as well as a ready to use decorator to specifically track deployments.